requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-ckeditor==6.7.1
Pillow==12.1.0
sqlparse==0.5.5
asgiref==3.11.1
tzdata==2025.3

